Refactors: Splitting up large files
