+++ truncate_client/src/app_inner.rs
@@ -1,36 +1,31 @@
-use eframe::egui::{self, Margin};
-use epaint::{vec2, Color32};
+use eframe::egui;
+use epaint::vec2;
 use instant::Duration;
 use truncate_core::{
     board::Board,
-    game::{self, GAME_COLORS, GAME_COLOR_BLUE, GAME_COLOR_RED},
-    generation::{self, generate_board, BoardSeed},
+    generation::{generate_board, BoardSeed},
     messages::{LobbyPlayerMessage, RoomCode, TruncateToken},
-    npc::scoring::{NPCParams, NPCPersonality},
+    npc::scoring::NPCPersonality,
 };
 
 use crate::{
+    handle_launch_code::handle_launch_code,
+    handle_messages::handle_server_msg,
     lil_bits::SplashUI,
     regions::{
-        active_game::{ActiveGame, GameLocation, HeaderType},
+        active_game::{ActiveGame, HeaderType},
         generator::GeneratorState,
         lobby::Lobby,
+        native_menu::render_native_menu_if_required,
         replayer::ReplayerState,
         single_player::SinglePlayerState,
         tutorial::TutorialState,
     },
-    utils::{
-        daily::{get_playable_daily_puzzle, get_puzzle_day, get_raw_daily_puzzle},
-        game_evals::get_main_dict,
-        macros::tr_log,
-        text::TextHelper,
-        urls::back_to_menu,
-        Lighten,
-    },
+    utils::{daily::get_puzzle_day, urls::back_to_menu},
 };
 
 use super::OuterApplication;
-use truncate_core::messages::{GameMessage, GameStateMessage, PlayerMessage};
+use truncate_core::messages::PlayerMessage;
 
 pub enum GameStatus {
     None(RoomCode, Option<TruncateToken>),
@@ -49,263 +44,8 @@ pub enum GameStatus {
     HardError(Vec<String>),
 }
 
-pub fn handle_server_msg(outer: &mut OuterApplication, ui: &mut egui::Ui, current_time: Duration) {
-    let mut recv = || match outer.rx_game.try_next() {
-        Ok(Some(msg)) => Ok(msg),
-        _ => Err(()),
-    };
-
-    while let Ok(msg) = recv() {
-        match msg {
-            GameMessage::Ping => {}
-            GameMessage::JoinedLobby(player_index, id, players, board, token) => {
-                // If we're already in a lobby, treat this as a lobby update
-                // (the websocket probably dropped and reconnected)
-                if let GameStatus::PendingStart(lobby) = &mut outer.game_status {
-                    if lobby.room_code.to_uppercase() == id.to_uppercase() {
-                        lobby.players = players;
-                        lobby.update_board(board, ui);
-                        continue;
-                    }
-                }
-
-                #[cfg(target_arch = "wasm32")]
-                {
-                    let local_storage =
-                        web_sys::window().unwrap().local_storage().unwrap().unwrap();
-                    local_storage
-                        .set_item("truncate_active_token", &token)
-                        .unwrap();
-
-                    // If we're joining a lobby, update the URL to match
-                    _ = web_sys::window()
-                        .unwrap()
-                        .location()
-                        .set_hash(id.to_uppercase().as_str());
-                }
-
-                outer.game_status = GameStatus::PendingStart(Lobby::new(
-                    ui.ctx(),
-                    id.to_uppercase(),
-                    players,
-                    player_index,
-                    board,
-                    outer.map_texture.clone(),
-                ))
-            }
-            GameMessage::LobbyUpdate(_player_index, _id, players, board) => {
-                match &mut outer.game_status {
-                    GameStatus::PendingStart(editor_state) => {
-                        // TODO: Assert that this message is for the correct lobby
-                        editor_state.players = players;
-                        editor_state.update_board(board, ui);
-                    }
-                    _ => panic!("Game update hit an unknown state"),
-                }
-            }
-            GameMessage::StartedGame(GameStateMessage {
-                room_code,
-                players,
-                player_number,
-                next_player_number,
-                board,
-                hand,
-                changes: _,
-            }) => {
-                // If we're already in a game, treat this as a game update
-                // (the websocket probably dropped and reconnected)
-                if let GameStatus::Active(game) = &mut outer.game_status {
-                    if game.depot.gameplay.room_code.to_uppercase() == room_code.to_uppercase() {
-                        let update = GameStateMessage {
-                            room_code,
-                            players,
-                            player_number,
-                            next_player_number,
-                            board,
-                            hand,
-                            changes: vec![], // TODO: Try get latest changes on reconnect without dupes
-                        };
-                        game.apply_new_state(update);
-                        continue;
-                    }
-                }
-
-                outer.game_status = GameStatus::Active(ActiveGame::new(
-                    ui.ctx(),
-                    room_code.to_uppercase(),
-                    None,
-                    None,
-                    players,
-                    player_number,
-                    next_player_number,
-                    board,
-                    hand,
-                    outer.map_texture.clone(),
-                    outer.theme.clone(),
-                    GameLocation::Online,
-                ));
-            }
-            GameMessage::GameUpdate(state_message) => match &mut outer.game_status {
-                GameStatus::Active(game) => {
-                    game.apply_new_state(state_message);
-                }
-                _ => todo!("Game update hit an unknown state"),
-            },
-            GameMessage::GameEnd(state_message, winner) => {
-                #[cfg(target_arch = "wasm32")]
-                {
-                    let local_storage =
-                        web_sys::window().unwrap().local_storage().unwrap().unwrap();
-                    local_storage.remove_item("truncate_active_token").unwrap();
-                }
-
-                match &mut outer.game_status {
-                    GameStatus::Active(game) => {
-                        game.apply_new_state(state_message);
-                        outer.game_status = GameStatus::Concluded(game.clone(), winner);
-                    }
-                    _ => {}
-                }
-            }
-            GameMessage::GameError(_id, _num, err) => match &mut outer.game_status {
-                GameStatus::Active(game) => {
-                    // assert_eq!(game.room_code, id);
-                    // assert_eq!(game.player_number, num);
-                    game.depot.gameplay.error_msg = Some(err);
-                }
-                _ => {}
-            },
-            GameMessage::GenericError(err) => {
-                outer.error = Some(err);
-            }
-            GameMessage::SupplyDefinitions(definitions) => {
-                match &mut outer.game_status {
-                    GameStatus::SinglePlayer(game) => {
-                        game.hydrate_meanings(definitions.clone());
-                        if let Some(dict_ui) = &mut game.active_game.dictionary_ui {
-                            dict_ui.load_definitions(definitions);
-                        }
-                    }
-                    GameStatus::Active(active_game) => {
-                        if let Some(dict_ui) = &mut active_game.dictionary_ui {
-                            dict_ui.load_definitions(definitions);
-                        }
-                    }
-                    _ => { /* Soft unreachable */ }
-                }
-            }
-            GameMessage::LoggedInAs(player_token) => {
-                #[cfg(target_arch = "wasm32")]
-                {
-                    let local_storage =
-                        web_sys::window().unwrap().local_storage().unwrap().unwrap();
-                    local_storage
-                        .set_item("truncate_player_token", &player_token)
-                        .unwrap();
-                }
-
-                outer.logged_in_as = Some(player_token);
-            }
-            GameMessage::ResumeDailyPuzzle(latest_puzzle_state, best_puzzle) => {
-                let mut puzzle_game = get_playable_daily_puzzle(
-                    ui.ctx(),
-                    latest_puzzle_state.puzzle_day,
-                    &outer.map_texture,
-                    &outer.theme,
-                    &outer.backchannel,
-                );
-
-                if let Some(best_puzzle) = best_puzzle {
-                    let mut best_game = puzzle_game.game.clone();
-                    best_game.rules.battle_delay = 0;
-                    let dict_lock = get_main_dict();
-                    let dict = dict_lock.as_ref().unwrap();
-
-                    for next_move in best_puzzle.current_moves.into_iter() {
-                        if best_game
-                            .play_turn(next_move, Some(dict), Some(dict), None)
-                            .is_err()
-                        {
-                            break;
-                        }
-                    }
-
-                    puzzle_game.best_game = Some(best_game);
-                }
-
-                let unplayed_puzzle = puzzle_game.clone();
-
-                match &mut puzzle_game.header {
-                    HeaderType::Summary { attempt, .. } => {
-                        *attempt = Some(latest_puzzle_state.attempt as usize)
-                    }
-                    _ => {}
-                }
-                puzzle_game.move_sequence = latest_puzzle_state.current_moves.clone();
-
-                let delay = puzzle_game.game.rules.battle_delay;
-                puzzle_game.game.rules.battle_delay = 0;
-                for next_move in latest_puzzle_state.current_moves.into_iter() {
-                    if puzzle_game
-                        .handle_move(next_move, &outer.backchannel)
-                        .is_err()
-                    {
-                        puzzle_game = unplayed_puzzle;
-                        break;
-                    }
-                }
-                puzzle_game.game.rules.battle_delay = delay;
-
-                puzzle_game.active_game.depot.ui_state.game_header = puzzle_game.header.clone();
-                outer.game_status = GameStatus::SinglePlayer(puzzle_game);
-            }
-            GameMessage::DailyStats(stats) => match &mut outer.game_status {
-                GameStatus::SinglePlayer(game) => {
-                    game.daily_stats = Some(stats);
-                }
-                _ => {}
-            },
-            GameMessage::LoadDailyReplay(puzzle_state) => {
-                let (seed, info) = get_raw_daily_puzzle(puzzle_state.puzzle_day);
-                let human_starts = info.as_ref().map(|(h, _)| *h).unwrap_or(true);
-
-                let mut game = game::Game::new(9, 9, Some(seed.seed as u64));
-                if human_starts {
-                    game.add_player("You".into());
-                    game.add_player("Computer".into());
-
-                    game.players[0].color = GAME_COLOR_BLUE;
-                    game.players[1].color = GAME_COLOR_RED;
-                } else {
-                    game.add_player("Computer".into());
-                    game.add_player("You".into());
-
-                    game.players[0].color = GAME_COLOR_RED;
-                    game.players[1].color = GAME_COLOR_BLUE;
-                }
-
-                let mut board = generation::generate_board(seed.clone())
-                    .expect("Common seeds should always generate a board")
-                    .board;
-                board.cache_special_squares();
-                game.board = board.clone();
-
-                let replayer = ReplayerState::new(
-                    ui.ctx(),
-                    outer.map_texture.clone(),
-                    outer.theme.clone(),
-                    game,
-                    puzzle_state.current_moves,
-                    if human_starts { 0 } else { 1 },
-                );
-                outer.game_status = GameStatus::Replay(replayer);
-            }
-        }
-    }
-}
-
 pub fn render(outer: &mut OuterApplication, ui: &mut egui::Ui, current_time: Duration) {
-    handle_server_msg(outer, ui, current_time);
+    handle_server_msg(outer, ui);
 
     if outer.log_frames {
         let ctx = ui.ctx().clone();
@@ -319,12 +59,9 @@ pub fn render(outer: &mut OuterApplication, ui: &mut egui::Ui, current_time: Dur
             });
     }
 
-    let mut send = |msg| {
-        outer.tx_player.try_send(msg).unwrap();
-    };
-
     // Block all further actions until we have a login token from the server,
     // or until the player accepts to play offline.
+    // TODO: Daily puzzle is still inaccessible offline, thus needs a similar check.
     if let (Some(waiting_for_login), None) = (&outer.started_login_at, &outer.logged_in_as) {
         if (current_time - *waiting_for_login) < Duration::from_secs(5) {
             SplashUI::new(vec!["INITIALIZING".to_string()])
@@ -357,303 +94,18 @@ pub fn render(outer: &mut OuterApplication, ui: &mut egui::Ui, current_time: Dur
     }
 
     let mut new_game_status = None;
-
-    if let Some(launched_room) = outer.launched_room.take() {
-        if launched_room == "__REJOIN__" {
-            match &mut outer.game_status {
-                GameStatus::None(_, Some(token)) => {
-                    send(PlayerMessage::RejoinGame(token.to_string()));
-                    new_game_status = Some(GameStatus::PendingJoin("...".into()));
-                }
-                _ => {
-                    new_game_status =
-                        Some(GameStatus::HardError(vec!["Could not rejoin".to_string()]));
-                }
-            }
-        } else if launched_room == "TUTORIAL_RULES" {
-            new_game_status = Some(GameStatus::Tutorial(TutorialState::new_rules(
-                ui.ctx(),
-                outer.map_texture.clone(),
-                &outer.theme,
-            )));
-            send(PlayerMessage::StartedTutorial {
-                name: "rules".to_string(),
-            });
-        } else if launched_room == "TUTORIAL_EXAMPLE" {
-            new_game_status = Some(GameStatus::Tutorial(TutorialState::new_example(
-                ui.ctx(),
-                outer.map_texture.clone(),
-                &outer.theme,
-            )));
-            send(PlayerMessage::StartedTutorial {
-                name: "example".to_string(),
-            });
-        } else if launched_room == "SINGLE_PLAYER" {
-            let mut board = Board::new(9, 9);
-            board.grow();
-            new_game_status = Some(GameStatus::PendingSinglePlayer(Lobby::new(
-                ui.ctx(),
-                "Single Player".into(),
-                vec![
-                    LobbyPlayerMessage {
-                        name: "You".into(),
-                        index: 0,
-                        color: (128, 128, 255),
-                    },
-                    LobbyPlayerMessage {
-                        name: "Computer".into(),
-                        index: 1,
-                        color: (255, 80, 80),
-                    },
-                ],
-                0,
-                board,
-                outer.map_texture.clone(),
-            )));
-            send(PlayerMessage::StartedSinglePlayer);
-        } else if launched_room == "DAILY_PUZZLE" {
-            let day = get_puzzle_day(current_time);
-            if let Some(token) = &outer.logged_in_as {
-                send(PlayerMessage::LoadDailyPuzzle(token.clone(), day));
-            }
-
-            new_game_status = Some(GameStatus::PendingDaily);
-        } else if launched_room == "RANDOM_PUZZLE" {
-            let seed = (current_time.as_micros() % 243985691) as u32;
-            let board_seed = BoardSeed::new(seed);
-            let board = generate_board(board_seed.clone())
-                .expect("Common seeds can be reasonably expected to produce a board")
-                .board;
-            let header = HeaderType::Summary {
-                title: format!("Regular Puzzle"),
-                attempt: None,
-            };
-            let puzzle_game = SinglePlayerState::new(
-                ui.ctx(),
-                outer.map_texture.clone(),
-                outer.theme.clone(),
-                board,
-                Some(board_seed),
-                true,
-                header,
-                NPCPersonality::jet(),
-            );
-            new_game_status = Some(GameStatus::SinglePlayer(puzzle_game));
-            send(PlayerMessage::StartedRandomPuzzle {
-                personality: "jet".into(),
-            });
-        } else if launched_room == "RANDOM_EASY_PUZZLE" {
-            let seed = (current_time.as_micros() % 243985691) as u32;
-            let board_seed = BoardSeed::new(seed);
-            let board = generate_board(board_seed.clone())
-                .expect("Common seeds can be reasonably expected to produce a board")
-                .board;
-            let header = HeaderType::Summary {
-                title: format!("Easy Puzzle"),
-                attempt: None,
-            };
-            let puzzle_game = SinglePlayerState::new(
-                ui.ctx(),
-                outer.map_texture.clone(),
-                outer.theme.clone(),
-                board,
-                Some(board_seed),
-                true,
-                header,
-                NPCPersonality::mellite(),
-            );
-            new_game_status = Some(GameStatus::SinglePlayer(puzzle_game));
-            send(PlayerMessage::StartedRandomPuzzle {
-                personality: "mellite".into(),
-            });
-        } else if launched_room.starts_with("PUZZLE:") {
-            let url_segments = launched_room.chars().filter(|c| *c == ':').count();
-            let has_board_generation = url_segments >= 3;
-            let has_npc_id = url_segments >= 4;
-
-            let mut parts = launched_room.split(':').skip(1);
-            let board_generation = if has_board_generation {
-                parts.next().map(str::parse::<u32>)
-            } else {
-                Some(Ok(0))
-            };
-            let npc = if has_npc_id {
-                parts
-                    .next()
-                    .map(|p| NPCPersonality::from_id(p.to_ascii_lowercase()))
-            } else {
-                Some(Some(NPCPersonality::jet()))
-            };
-            let seed = parts.next().map(str::parse::<u32>);
-            let player = parts
-                .next()
-                .map(|p| p.parse::<usize>().unwrap_or(0))
-                .unwrap_or(0);
-
-            if let (Some(Ok(board_generation)), Some(Some(npc)), Some(Ok(seed))) =
-                (board_generation, npc, seed)
-            {
-                let board_seed = BoardSeed::new_with_generation(board_generation, seed);
-                let board = generate_board(board_seed.clone())
-                    .expect("Common seeds can be reasonably expected to produce a board")
-                    .board;
-                let header = HeaderType::Summary {
-                    title: format!("Truncate Puzzle"),
-                    attempt: None,
-                };
-                send(PlayerMessage::StartedRandomPuzzle {
-                    personality: npc.name.clone(),
-                });
-                let puzzle_game = SinglePlayerState::new(
-                    ui.ctx(),
-                    outer.map_texture.clone(),
-                    outer.theme.clone(),
-                    board,
-                    Some(board_seed),
-                    player == 0,
-                    header,
-                    npc,
-                );
-                new_game_status = Some(GameStatus::SinglePlayer(puzzle_game));
-            } else {
-                new_game_status = Some(GameStatus::HardError(vec![
-                    "Sorry, that puzzle URL".to_string(),
-                    "doesn't look right!".to_string(),
-                ]));
-            }
-        } else if launched_room.starts_with("REPLAY:") {
-            if let Some(id) = launched_room.split(':').skip(1).next() {
-                send(PlayerMessage::LoadReplay(id.to_string()));
-                new_game_status = Some(GameStatus::PendingReplay);
-            } else {
-                new_game_status = Some(GameStatus::HardError(vec![
-                    "Sorry, that replay URL".to_string(),
-                    "doesn't look right!".to_string(),
-                ]));
-            }
-        } else if launched_room == "DEBUG_BEHEMOTH" {
-            let behemoth_board = Board::from_string(include_str!("../tutorials/test_board.txt"));
-            let seed_for_hand_tiles = BoardSeed::new_with_generation(0, 1);
-            let behemoth_game = SinglePlayerState::new(
-                ui.ctx(),
-                outer.map_texture.clone(),
-                outer.theme.clone(),
-                behemoth_board,
-                Some(seed_for_hand_tiles),
-                true,
-                HeaderType::Timers,
-                NPCPersonality::jet(),
-            );
-            new_game_status = Some(GameStatus::SinglePlayer(behemoth_game));
-            outer.log_frames = true;
-        } else if launched_room.is_empty() {
-            // No room code means we start a new game.
-            send(PlayerMessage::NewGame(outer.name.clone()));
-            new_game_status = Some(GameStatus::PendingCreate);
-        } else {
-            let token = if let GameStatus::None(_, token) = &outer.game_status {
-                token.clone()
-            } else {
-                None
-            };
-
-            send(PlayerMessage::JoinGame(
-                launched_room.clone(),
-                outer.name.clone(),
-                token,
-            ));
-
-            new_game_status = Some(GameStatus::PendingJoin(launched_room));
-        }
+    if let Some(launched_code) = outer.launched_code.take() {
+        new_game_status = handle_launch_code(&launched_code, outer, ui);
     }
 
+    render_native_menu_if_required(outer, ui);
+
+    let mut send = |msg| {
+        outer.tx_player.try_send(msg).unwrap();
+    };
+
     match &mut outer.game_status {
-        GameStatus::None(room_code, token) => {
-            if ui.button("Generator").clicked() {
-                new_game_status = Some(GameStatus::Generator(GeneratorState::new(
-                    ui.ctx(),
-                    outer.map_texture.clone(),
-                    outer.theme.clone(),
-                )));
-            }
-            if ui.button("Tutorial: Rules").clicked() {
-                new_game_status = Some(GameStatus::Tutorial(TutorialState::new_rules(
-                    ui.ctx(),
-                    outer.map_texture.clone(),
-                    &outer.theme,
-                )));
-            }
-            if ui.button("Tutorial: Example").clicked() {
-                new_game_status = Some(GameStatus::Tutorial(TutorialState::new_example(
-                    ui.ctx(),
-                    outer.map_texture.clone(),
-                    &outer.theme,
-                )));
-            }
-            if ui.button("Single Player").clicked() {
-                let mut board = Board::new(9, 9);
-                board.grow();
-                new_game_status = Some(GameStatus::PendingSinglePlayer(Lobby::new(
-                    ui.ctx(),
-                    "Single Player".into(),
-                    vec![
-                        LobbyPlayerMessage {
-                            name: "You".into(),
-                            index: 0,
-                            color: (128, 128, 255),
-                        },
-                        LobbyPlayerMessage {
-                            name: "Computer".into(),
-                            index: 1,
-                            color: (255, 80, 80),
-                        },
-                    ],
-                    0,
-                    board,
-                    outer.map_texture.clone(),
-                )));
-            }
-            if ui.button("Behemoth").clicked() {
-                let behemoth_board =
-                    Board::from_string(include_str!("../tutorials/test_board.txt"));
-                let seed_for_hand_tiles = BoardSeed::new_with_generation(0, 1);
-                let behemoth_game = SinglePlayerState::new(
-                    ui.ctx(),
-                    outer.map_texture.clone(),
-                    outer.theme.clone(),
-                    behemoth_board,
-                    Some(seed_for_hand_tiles),
-                    true,
-                    HeaderType::Timers,
-                    NPCPersonality::jet(),
-                );
-                new_game_status = Some(GameStatus::SinglePlayer(behemoth_game));
-                outer.log_frames = true;
-            }
-            if ui.button("New Game").clicked() {
-                // TODO: Send player name in NewGame message
-                send(PlayerMessage::NewGame(outer.name.clone()));
-                new_game_status = Some(GameStatus::PendingCreate);
-            }
-            ui.horizontal(|ui| {
-                ui.text_edit_singleline(room_code);
-                if ui.button("Join Game").clicked() {
-                    send(PlayerMessage::JoinGame(
-                        room_code.clone(),
-                        outer.name.clone(),
-                        token.clone(),
-                    ));
-                    new_game_status = Some(GameStatus::PendingJoin(room_code.clone()));
-                }
-            });
-            if let Some(existing_token) = token {
-                ui.label("Existing game found, would you like to rejoin?");
-                if ui.button("Rejoin").clicked() {
-                    send(PlayerMessage::RejoinGame(existing_token.to_string()));
-                    new_game_status = Some(GameStatus::PendingJoin("...".into()));
-                }
-            }
-        }
+        GameStatus::None(_, _) => { /* handled above */ }
         GameStatus::Generator(generator) => {
             generator.render(ui, &outer.theme, current_time);
         }
@@ -786,12 +238,7 @@ pub fn render(outer: &mut OuterApplication, ui: &mut egui::Ui, current_time: Dur
 
             let resp = splash.render(ui, &outer.theme, current_time, &outer.map_texture);
             if matches!(resp.clicked, Some("reload")) {
-                // TODO: Neatly kick back to the wrapper page without a reload
-                #[cfg(target_arch = "wasm32")]
-                {
-                    _ = web_sys::window().unwrap().location().set_hash("");
-                    _ = web_sys::window().unwrap().location().reload();
-                }
+                back_to_menu();
             }
         }
     }
