+++ truncate_client/src/regions/mod.rs
@@ -1,6 +1,7 @@
 pub mod active_game;
 pub mod generator;
 pub mod lobby;
+pub mod native_menu;
 pub mod replayer;
 pub mod single_player;
 pub mod tutorial;
